scrapper/requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
cssselect==1.2.0
fake-useragent==1.5.1
idna==3.7
importlib_metadata==8.2.0
lxml==5.2.2
lxml_html_clean==0.1.1
parse==1.20.2
pyee==11.1.0
pyppeteer==2.0.0
pyquery==2.0.0
requests==2.32.3
requests-html==0.10.0
soupsieve==2.5
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==1.26.19
w3lib==2.2.1
websockets==10.4
zipp==3.19.2
pillow==10.4.0